Modified these files for effective navigation
