+++ app/src/main/AndroidManifest.xml
@@ -14,80 +14,16 @@
         android:theme="@style/Theme.Disheecompose"
         tools:targetApi="31">
         <activity
-            android:name=".AddItem"
-            android:exported="false"
-            android:label="@string/title_activity_add_item"
-            android:theme="@style/Theme.Disheecompose">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".Cart"
-            android:exported="false"
-            android:label="@string/title_activity_cart"
-            android:theme="@style/Theme.Disheecompose">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".Success"
-            android:exported="false"
-            android:label="@string/title_activity_success"
-            android:theme="@style/Theme.Disheecompose">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".Welcome"
-            android:exported="false"
-            android:label="@string/title_activity_welcome"
-            android:theme="@style/Theme.Disheecompose">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".Feed"
-            android:exported="false"
-            android:label="@string/title_activity_feed"
-            android:theme="@style/Theme.Disheecompose">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
-            android:name=".UploadPicture"
-            android:exported="false"
-            android:label="@string/title_activity_upload_picture"
-            android:theme="@style/Theme.Disheecompose" />
-        <activity
-            android:name=".Login"
+            android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/title_activity_login"
+            android:label="@string/app_name"
             android:theme="@style/Theme.Disheecompose">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
+                <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity
-            android:name=".Home"
-            android:exported="true"
-            android:label="@string/title_activity_home"
-            android:theme="@style/Theme.Disheecompose" />
-        <activity
-            android:name=".PostDetails"
-            android:exported="false"
-            android:label="@string/title_activity_post_details"
-            android:theme="@style/Theme.Disheecompose" />
-        <activity
-            android:name=".MainActivity"
-            android:exported="true"
-            android:theme="@style/Theme.Disheecompose" />
     </application>
 
 </manifest>
